python/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
decorator==4.4.2
idna==3.4
imageio==2.25.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.24.2
Pillow==9.4.0
proglog==0.1.10
pytube==12.1.2
requests==2.28.2
tqdm==4.64.1
urllib3==1.26.14
